pyproject.toml
[tool.poetry]
name = "sparkjobtest"
version = "0.1.10"
description = ""
authors = ["Col Perks <wild.fauve@gmail.com>"]

[tool.poetry.plugins."job_entry_point_group"]
job_main = "sparkjobtest.job:job"


[tool.poetry.dependencies]
python = "^3.9"
pyspark = "3.2.0"
pendulum = "^2.1.2"
simple-memory-cache = "^1.0.0"
delta-spark = "^2.0.0"
PyMonad = "^2.4.0"
rdflib = "^6.2.0"
rdflib-jsonld = "^0.6.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
